Extract manifest in dedicated file
